added two placement helper functions to place_utils.cpp/.h
